First attempt at new UPDATE mode for DC1 phoSim Pass II

+++ NERSC/setupVisit.py
@@ -44,13 +44,17 @@ log.info('PhoSim output directory = \n\t%s',outDir)
 
 ### Is there a previous instance of this job step that requires archiving?
 
-if not os.access(outDir,os.F_OK):                          ## NO
+### But FIRST, if "UPDATE" mode, just check that archive dir: 1)
+### exists => do nothing; 2) not exists => create it (like a fresh
+### visit)
+
+if not os.access(outDir,os.F_OK):                                      ## NO
     log.info('Creating output directory: \n\t%s',outDir)
     os.makedirs(outDir,filePermissions)  ## Note that filePermissions is ignored on cori
 #    os.chmod(outDir,filePermissions)
     os.system('chmod -R 0755 '+outDir)  ## may be obsolete now we are using ACLs
     pass
-else:                                                      ## YES
+elif os.getenv('DC1_SETUP_MODE') == 'NORMAL':          ## YES
     ## Create appropriate archive directory, 'archive-N'
     archiveRoot = os.path.join(outDir,archivesDirName)
     if not os.access(archiveRoot,os.F_OK):
@@ -108,18 +112,20 @@ if rc <> 0 :
 ## Check that SCRATCH (staging) area is clean
 scrDir = os.path.join(os.getenv('PHOSIM_SCR_ROOT'),os.getenv('DC1_SIXDIGSTREAM'))
 log.info('Checking phoSim scratch/staging space: '+scrDir)
-if os.access(scrDir,os.F_OK):
+scr_work=os.path.join(scrDir,'work')
+scr_output=os.path.join(scrDir,'output')
+if os.access(scrDir,os.F_OK) and os.getenv('DC1_SETUP_MODE') == 'NORMAL':
     log.info('phoSim scratch/staging directory already exists.  Cleaning up...')
     shutil.rmtree(scrDir)
     pass
 
-
 ## Create staging area (work and output directories) in SCRATCH
-log.info('Creating phoSim work and output directories in $SCRATCH')
-scr_work=os.path.join(scrDir,'work')
-scr_output=os.path.join(scrDir,'output')
-os.makedirs(scr_work,filePermissions)
-os.makedirs(scr_output,filePermissions)
+if os.getenv('DC1_SETUP_MODE') == 'NORMAL':
+    log.info('Creating phoSim work and output directories in $SCRATCH')
+    os.makedirs(scr_work,filePermissions)
+    os.makedirs(scr_output,filePermissions)
+    pass
+
 
 cmd = 'pipelineSet DC1_SCR_PHOSIMOUT '+scrDir
 print cmd
@@ -136,10 +142,32 @@ if rc <> 0 :
 stream = int(os.getenv('PIPELINE_STREAM'))
 visitFile = os.getenv('DC1_VISIT_DB')
 log.info('Extract visit data.')
-(visitID,sensorList) = getVisit(stream,visitFile)
+(visitID,fullSensorList) = getVisit(stream,visitFile)
 visitID = str(visitID)
 print 'visitID = ',visitID,', type(visitID) = ',type(visitID)
-print 'sensorList = ',sensorList
+print 'fullSensorList = ',fullSensorList
+sensorList = fullSensorList
+
+#########################################
+##### "UPDATE" mode for revised visitDB
+#########################################
+if os.getenv('DC1_SETUP_MODE') == 'UPDATE':
+    log.info('Running in UPDATE mode')
+    visitFile2 = os.getenv('DC1_VISIT_DB2')
+    (visitID2,sensorList2) = getVisit(stream,visitFile2)
+    if visitID != visitID2:
+        log.error('Obtained different visitIDs for same stream')
+        sys.exit(1)
+        pass
+    s1 = set(fullSensorList)
+    s2 = set(sensorList2)
+    sinterx = s2 & s1
+    newSensorList = list(s2 - sinterx)
+    print 'New sensor list [',len(newSensorList),'] : ',newSensorList
+    sensorList = newSensorList
+    pass
+                  
+
 
 ## Package up the sensorList into variables of <990 characters each (email limitation)
 ### Figure 8 characters/sensor, e.g. R01_S12 plus a comma
@@ -167,14 +195,15 @@ if rc1 <> 0 or rc2 <> 0 or rc3 <> 0:
 
 ##  Set up phoSim instanceCatalog
 
-if os.getenv('PHOSIM_IC_GEN') == 'STATIC':
+icName = 'phosim_cat_'+visitID+'.txt'
+print 'icName = ',icName
+icSelect = os.path.join(icDir,icName)
+print 'icSelect = ',icSelect
+
+    if os.getenv('PHOSIM_IC_GEN') == 'STATIC':
     print 'Using statically generated instance catalog'
     icDir = os.getenv('PHOSIM_CATALOGS')+'/'+visitID
     print 'icDir = ',icDir
-    icName = 'phosim_cat_'+visitID+'.txt'
-    print 'icName = ',icName
-    icSelect = os.path.join(icDir,icName)
-    print 'icSelect = ',icSelect
     
 elif os.getenv('PHOSIM_IC_GEN') == 'DYNAMIC':
     print 'Using dymanically generated instance catalog'
@@ -183,10 +212,6 @@ elif os.getenv('PHOSIM_IC_GEN') == 'DYNAMIC':
     print 'cmd = ',cmd
     icDir = os.path.join(scrDir,'instCat')
     print 'icDir = ',icDir
-    icName = 'phosim_cat_'+visitID+'.txt'
-    print 'icName = ',icName
-    icSelect = os.path.join(icDir,icName)
-    print 'icSelect = ',icSelect
     minMag = os.getenv('DC1_MINMAG')
     print 'minMag = ',minMag
     opts = ' --db '+os.getenv('DC1_OPSIM_DB')+' --out '+icDir+' --id '+visitID+' --min_mag '+minMag
@@ -202,8 +227,14 @@ elif os.getenv('PHOSIM_IC_GEN') == 'DYNAMIC':
         log.error('Failed to generate instance catalog.')
         sys.exit(1)
         pass
-    pass
+elif os.getenv('PHOSIM_IC_GEN') == 'EXISTING':
+    print 'Using existing instanceCatalog'
+    icDir = os.path.join(scrDir,'instCat')
+    print 'icDir = ',icDir
 
+## Should probably check to see if this directory really exists!
+    
+    pass
 
 
 ## Preserve the instance catalog info for subsequent processing steps
